# stms-backend/openapi.yaml
openapi: 3.0.3
info:
  title: STMS - Smart Trainer Management System API
  version: 1.0.0
  description: |
    Authentication & User Management API for Sprint 1.
    Multi-tenant athletics management platform with club-scoped data isolation.
  contact:
    name: STMS Team
    email: dev@stms.example.com
  license:
    name: Proprietary
servers:
  - url: https://api.stms.example.com/v1
    description: Production
  - url: https://api-staging.stms.example.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RS256 signed JWT access token (15 min expiry).
        Include in Authorization header: `Bearer <access_token>`
    RefreshTokenCookie:
      type: apiKey
      in: cookie
      name: refreshToken
      description: HttpOnly, Secure, SameSite=Strict refresh token cookie (7 days)
  schemas:
...
    Error:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Field-level validation errors (400 only)
    Success:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
...
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: "coach.john@aditya.edu"
        password:
          type: string
          minLength: 12
          maxLength: 128
          description: NIST 2025 guidelines - 12+ chars, no composition rules
          example: "correct-horse-battery-staple"
        role:
          type: string
          enum: [athlete, coach]
          description: Initial role; club_admin/system_admin created via invitation
          example: "coach"
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "John Smith"
        clubId:
          type: string
          description: Optional - join existing club on registration (requires invite token)
          example: "507f1f77bcf86cd799439011"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    GoogleAuthRequest:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
          description: Google OAuth ID token from frontend
        clubId:
          type: string
          description: Optional club to join
    TokenResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            accessToken:
              type: string
              description: RS256 JWT, 15 min expiry
            expiresIn:
              type: integer
              description: Seconds until expiry (900)
              example: 900
            tokenType:
              type: string
              enum: [Bearer]
              example: "Bearer"
            user:
              $ref: '#/components/schemas/UserResponse'
    RefreshResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            accessToken:
              type: string
            expiresIn:
              type: integer
              example: 900
            tokenType:
              type: string
              enum: [Bearer]
    AuthMeResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/UserResponse'
            permissions:
              type: array
              items:
                type: string
              example: ["attendance:read", "attendance:write", "workout:read"]
    SwitchClubRequest:
      type: object
      required: [clubId]
      properties:
        clubId:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          description: Must be in user's clubIds

    # ==================== USER ====================
    UserResponse:
      type: object
      properties:
        id:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
        firebaseUid:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [system_admin, club_admin, coach, athlete]
        clubIds:
          type: array
          items:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
        activeClubId:
          type: string
          nullable: true
          pattern: '^[0-9a-fA-F]{24}$'
        status:
          type: string
          enum: [active, invited, deactivated, deleted]
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user (email/password)
      description: |
        Creates Firebase Auth user + MongoDB user document.
        Returns access token + sets refresh token cookie.
        Role must be 'athlete' or 'coach'. Club admins created via invitation.
      security: []  # Public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited (5 requests per 15 min)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with email/password
      description: Returns access token + sets refresh token cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited (5 requests per 15 min)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/google:
    post:
      tags: [Authentication]
      summary: Login/register with Google OAuth
      description: Verifies Google ID token, creates/updates user, returns tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid ID token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Google token verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Uses refresh token from HttpOnly cookie. Rotates refresh token.
      security: []
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Refresh token invalid/expired/revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout - revokes refresh token
      description: Blacklists refresh token in Redis, clears cookie
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user with permissions
      description: Returns user profile, active club, and computed permissions
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          description: Invalid/expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/switch-club:
    post:
      tags: [Authentication]
      summary: Switch active club context
      description: Sets user's activeClubId (must be in their clubIds). Subsequent requests scoped to new club.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwitchClubRequest'
      responses:
        '200':
          description: Club switched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '400':
          description: Club not in user's memberships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Membership not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== USERS ====================
...